Fixed handling of oldPrice attribute
