+++ controllers/products/_price_table_modal.htm
@@ -46,6 +46,7 @@
     $(function () {
         var priceTable;
         // Give the widget time to render
+
         setTimeout(function () {
             priceTable = {
                 state: JSON.parse('<?= addslashes($pricetableState) ?>'),
